Change int to float

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
-"""This module contains a coroutine
+"""
+This module contains a coroutine
 """
 import asyncio
 from typing import List
@@ -7,7 +8,7 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> List[int]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """Spawns wait_random n times with the specified max_delay.
     """
     delays = []
